setup roomInfo route code
