+++ serverless-aws-sam/events/restAPIEvent.json
@@ -13,13 +13,13 @@
     ]
   },
   "pathParameters": {
-    "proxy": "/path"
+    "proxy": "path"
   },
   "stageVariables": {
     "baz": "qux"
   },
   "headers": {
-    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
+    "Accept": "text/html,application/json,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
     "Content-Type": "application/json",
     "Accept-Encoding": "gzip, deflate, sdch",
     "Accept-Language": "en-US,en;q=0.8",
