+++ serverless-aws-sam/src/ejs-page-render/roomInfo/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "include": ["../../types/**/*.ts", "./**/*.ts"],
+  "exclude": ["node_modules"],
+  "compilerOptions": {}
+}
